Creating an IntelliJ Idea project. Removing all the .class files. Adding a package name.
