--- SortVisualizer/bin/.gitignore
@@ -1,13 +1,0 @@
-/Colors.class
-/InsertionSortThread.class
-/MainWindow$1.class
-/MainWindow$ButtonListener.class
-/MainWindow$SliderListener.class
-/MainWindow$State.class
-/MainWindow.class
-/MergeSortThread.class
-/QuickSortThread.class
-/SelectionSortThread.class
-/SortPanel.class
-/SortThread$1.class
-/SortThread.class
