+++ SortVisualizer/src/main/java/com/algorithms/sortingvisualizer/MergeSortThread.java
@@ -1,3 +1,5 @@
+package com.algorithms.sortingvisualizer;
+
 import java.util.ArrayList;
